added a way to calculate every color for a palette
instead of giving an image, this allows the user to write "fillPal" to denote that they want to calculate what color that palette will replace every pixel color with.
the code is repeated in a gross way (should be calling a method with some parameters) and the resulting filled palettes are not particularly useful. they're huge (225ish MB in my experience) and they take a long time to open and close, meaning they don't really lead to much of a speed increase (thus why this is not on the main branch)